* Reverted tile stacking, seem to cause issues with id: 4641 and such on bigger maps
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@4014 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ thing.h
@@ -134,6 +134,6 @@ class Thing
 
 	private:
 		Cylinder* parent;
-		int32_t refCount;
+		int16_t refCount;
 };
 #endif
